chat-app/services/auth-service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
passlib
python-jose
python-multipart
pydantic
websockets
bcrypt